requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
websockets==11.0.3
anthropic==0.86.0
python-dotenv==1.2.1
